// src/__generated__/logsTableRelay_query.graphql.ts
export type logsTableRelay_query$data = {
  readonly logs: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly deviceId: string | null | undefined;
        readonly eventType: string;
        readonly hostname: string | null | undefined;
        readonly id: string;
        readonly ingestDay: string;
        readonly nickname: string | null | undefined;
        readonly organizationId: string | null | undefined;
        readonly organizationName: string | null | undefined;
        readonly severity: string;
        readonly summary: string | null | undefined;
        readonly timestamp: Instant;
        readonly toolEventId: string;
        readonly toolType: string;
        readonly " $fragmentSpreads": FragmentRefs<"logCopyButton_log" | "logDrawerDetails_log">;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "logsTableRelay_query";
};
export type logsTableRelay_query$key = {
  readonly " $data"?: logsTableRelay_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"logsTableRelay_query">;
};